fix(Date Range Picker): prevent drag-and-drop issue on first selected date
When attempting to drag and drop the first selected date onto itself caused the second selected date to be the same Date as the first selected date. This was due to handleClick event handler, which changes first/second selected date, firing every time you drag and drop the date onto itself.

Add condition to handleClick function that checks if date was dragged and dropped onto itself and exits handleClick, if it was.