fix: prevent shadow selected date from disappearing when dragging date
When dragging the selected date to the position where it exchanges with
another selected date, shadow selected date disappeares. For example:
dragging 27 onto 8, when another date is 10. This was due to useEffect
that was watching for draggedDate (first/second), and setting the shadow
to the first selected date when draggedDate was First, and to the second
selected date when draggedDate was Second. When draggedDate (first/second)
changed, the useEffct changed the shadowSelected date accordingly.

Introduce a variable isDragging, which evaluates to true when user starts
dragging and evaluates to false when they stop, make this useEffect watch
for this varible and only when the user completely stops dragging, remove
the shadow date from the screen.